Refactor code.
