+++ src/main/java/org/bigtows/window/ui/notetree/NoteTree.java
@@ -8,6 +8,7 @@ import org.bigtows.window.ui.notetree.tree.entity.Task;
 import org.bigtows.window.ui.notetree.tree.node.NoteTreeNode;
 import org.bigtows.window.ui.notetree.tree.node.TaskTreeNode;
 import org.bigtows.window.ui.notetree.tree.render.PinNoteTreeCellRender;
+import org.bigtows.window.ui.notetree.utils.ExpandTreeUtils;
 
 import javax.swing.*;
 import javax.swing.tree.*;
@@ -47,57 +48,15 @@ public class NoteTree extends JPanel {
     public void updateModel(List<MutableTreeNode> data) {
 
         var iterator = tree.getExpandedDescendants(new TreePath(tree.getModel().getRoot())).asIterator();
-        List<TreePath> context = new ArrayList<>();
+        List<TreePath> listOfLeaf = new ArrayList<>();
 
-        iterator.forEachRemaining(context::add);
+        iterator.forEachRemaining(listOfLeaf::add);
         SwingUtilities.invokeLater(() -> {
             tree.setModel(this.buildTreeModelByListTreeNode(data));
-            expandTree(tree, context);
+            ExpandTreeUtils.expandLeaf(tree, listOfLeaf);
         });
     }
 
-    private void expandTree(JTree tree, List<TreePath> context) {
-        TreeNode root = (TreeNode) tree.getModel().getRoot();
-        expandAll(tree, new TreePath(root), context);
-    }
-
-    private void expandAll(JTree tree, TreePath path, List<TreePath> context) {
-        TreeNode node = (TreeNode) path.getLastPathComponent();
-
-        if (node.getChildCount() >= 0) {
-            Enumeration enumeration = node.children();
-            while (enumeration.hasMoreElements()) {
-                TreeNode n = (TreeNode) enumeration.nextElement();
-                TreePath p = path.pathByAddingChild(n);
-
-                expandAll(tree, p, context);
-            }
-        }
-
-        if (this.findContext(context, path)) {
-            tree.expandPath(path);
-        }
-    }
-
-    private boolean findContext(List<TreePath> context, TreePath currentPath) {
-        var lastElementPath = currentPath.getLastPathComponent();
-        for (TreePath treePath : context) {
-            var treePathLast = treePath.getLastPathComponent();
-
-            if (lastElementPath instanceof NoteTreeNode && treePathLast instanceof NoteTreeNode) {
-                if (((NoteTreeNode) lastElementPath).getUserObject().getIdentity().equals(
-                        ((NoteTreeNode) treePathLast).getUserObject().getIdentity())) {
-                    return true;
-                }
-            } else if (lastElementPath instanceof TaskTreeNode && treePathLast instanceof TaskTreeNode) {
-                if (((TaskTreeNode) lastElementPath).getUserObject().getIdentity().equals(
-                        ((TaskTreeNode) treePathLast).getUserObject().getIdentity())) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
 
 
     private TreeModel buildTreeModelByListTreeNode(List<MutableTreeNode> mutableTreeNodes) {
